Added web-mode-twig-font-lock-keywords to fix highlighting problem.
For some reason the keywords were being set to
web-mode-uel-font-lock-keywords when encountering '{{' tags.  As far
as I can tell, UEL does not use '{{' at all.  This was causing
'}' (used by twig in array definitions) to be erroneously highlighted.

Then web-mode-django-font-lock-keywords were not completely
appropriate either.  This new set of keywords is a better fit for twig
and provides a place for further customization.