in go templates, check keywords correctly
Go templates allow defining your own functions, so that there could be a
function "endTheWorld", which would match "^end" incorrectly.

Two new functions, web-mode-block and web-mode-block-matches were added.
web-mode-block returns the current block as a string and
web-mode-block-matches checks a regexp against that block.

web-mode-block-matches would not be necessary if
web-mode-block-starts-with and web-mode-block-ends-with were adjusted to
check against web-mode-block, instead of using looking-at/looking-back
on the buffer, which prevents one from using something
like (web-mode-block-starts-with "^end$").