Fix error introduced in commit 0583569
Upon running with the current master, my Emacs (23.3.1) complains about 3
arguments given to the >= operator.

bisecting revelas commit 0583569 as the culpit.

To be completly honest, I'm not sure that this is the correct fix. However with
this fix, the all tests works again, and web-mode appears to be working as
expected again.