web-mode-colorize: re-enable css rgba() colorization
This is already selected out of the in the regex (/rgba?/), but it
falls to the default case in `web-mode-colorize` and doesn't get
colored. Instead, put it in the same case as rgb().

As was true in the original commit enabling this (`a04b31c`), the alpha
value is ignored.