web-mode-part-scan: faster JavaScript string parsing
Big JavaScript files can take a long time to open and edit:

    for i in $(seq 10000); do echo '"xxxxxxxxx"' >> test.js; done

    (setq start (float-time))
    (find-file "test.js")
    (setq end (float-time))
    (- end start)
    => 7.90

Profiling shows most time is spent in `looking-back` within
`web-mode-part-scan`. This is run once per string:

    'foo bar \\\' ack'
                ^
                |
     (looking-back "\\\\+'" reg-beg)

The documentation for `looking-back` stresses is slowness:

    As a general recommendation, try to avoid using `looking-back'
    wherever possible, since it is slow.

This commit circumvents `looking-back` by hand-coding its regex in
elisp. We're only looking for `\+`, so it's not unreasable to search for
it with a loop.

Results:

    (setq start (float-time))
    (find-file "test.js")
    (setq end (float-time))
    (- end start)
    => 0.28