web-mode-colorize: don't highlight [#fea]ture
Before, this CSS:

    #features {
    }

...would have `#fea` highlighted a yellowish color, since `#fea` is
a hex color. Personally I've encountered `#feedba`ck, and there's
hundreds of other English words that start with `[a-f]{3}`.

This commit, extends the regex to end with `[^[:alnum:]]`. This won't
damage existing highlighting (`#fffx` is invalid), and will throw out
English words with 4 or more characters.

This does not 100% fix CSS colorization; we'll continue to erroneously
colorize this CSS:

    #fea {
    }

...but fixing that would require a much bigger change.

Fixes #818.