Finally, add indentation; basically, if the previous tags block didn't properly close its first parenthesis, indent the newline an additional indent.
Unfortunately, – without look-behind/ahead – there's no real way to
check properly closed parentheses with Emacs's regex (much to my
chagrin) so I had to result to outputting the text in question to a
temp. buffer and checking said text with `forward-sexp`; it's not
ideal nor the cleanest but it works.