fix uncomment for tsx and jsx
For uncommenting, as far as I understand, we rely on some text properties with
value as 'comment, to determine the beginning of a comment. But this won't work
well for jsx elements ({/* ... */}). Because `web-mode-part-scan` only marks
'comment property starting from `/` thus missing `{`.

This PR makes sure `web-mode-part-scan` to include `{` in commenting, so
uncommenting will work. This is however not a complete fix but it would cover
most of the use cases.