fixed: web-mode-comment-indent-new-line
Like the behavior of the original `comment-indent-new-line` it replaces, it inserts a space when generating a new comment line.
I thought the behavior before the modification was a matter of preference and that I would have to provide a custom one.
However, when I actually looked at the source code, I found that it seemed meaningless to bother to concat an empty string.
Therefore, I decided that I simply forgot to insert a space.